add generic enumerator, enhance ballot and manifest generators
configure the ballot and manifest generators to support different modes of operation. Now you can provide an rng, a function to resolve votes, or an explicit list of votes when creating a fake ballot. Added a generic enumerator for the c++ collections that can be references from any class with enumerable properties.