create reassignment in elementmodp instance functions
adda TODO for ISSUE #189 and remove redundant native interface facade. Add ThrowIfInvalid to the electiongaurd safe handle as a convenience function for instance allocations using the new reassignment. Only execute reassignment if we pass the exception checks so that the caller has the option to handle the exception gracefully and still access the pre-existing value.