Add IElectionGuardian and use in most places
Add an interface for the IElectionGuardian so that different parts of the system can work with guardians without having ot know the object structure. Add the _commitmentSeed to the guardian which is used for seeding challenge responses during decryption. Move some of the structures like ElectionPublicKey to where they are more readily accessible.