Feature/verifiable decryption (WIP) (#233)
* Move hash functions from group to their own class

* move polynomial

* move base classes to a folder

* Rework c# interfaces for contests and selections

* move extensions

* Add IElectionGuardian and use in most places

Add an interface for the IElectionGuardian so that different parts of the system can work with guardians without having ot know the object structure. Add the _commitmentSeed to the guardian which is used for seeding challenge responses during decryption. Move some of the structures like ElectionPublicKey to where they are more readily accessible.

* adjust the cp proof stub

* tests reafactor

* rename refactor shares classes

* split out decryption and accumulation

previously it would do both in the same step, now we separate out the accumulation so it can be used to generate proofs

* Add challenges

* add challenge responses

* Add CP Proof implementation

* deterministic fake manifest

make the generation of test data deterministic for comparisons

* rework the challenge responses

* consistent documentation and disposal patterns

* null coalesce during dispsal of native objects

* fix xargs copy

* use bigmath for intermediate value

when computing the share, we need to use bigmath for the interpolated value so that we are not writing to the share object

* add equality comparison to ciphertext

add more equations to comments

* add CommitmentOffset

* use CommitmentOffset when validating challenges

* refactor decryption tests

* Address PR feedback

* fix compilation errors

* address PR feedback

* pr fedback

* pr feedback

* add convenience api for adding ballots